+++ lib/generators/test_data/environment_file_generator.rb
@@ -9,6 +9,10 @@ module TestData
         require_relative "development"
 
         Rails.application.configure do
+          # Rails creates secret key base for only "development" and "test"
+          #   For more info, see: https://github.com/testdouble/test_data/issues/2
+          self.secrets.secret_key_base ||= Rails.application.send(:generate_development_secret)
+
           # Don't persist schema.rb or structure.sql after test_data is migrated
           config.active_record.dump_schema_after_migration = false
         end